MISC: default paths when not existing yet
On fresh systems, ~/.pcvs & $PREFIX/.pcvs-build may not exist yet,
leading to exceptions when trying to save file in their tree.

+++ pcvs/helpers/utils.py
@@ -90,6 +90,8 @@ def set_local_path(path):
 ####################################
 
 def create_or_clean_path(prefix):
+    if not os.path.exists(prefix):
+        return
     if os.path.isdir(prefix):
         shutil.rmtree(prefix)
         os.mkdir(prefix)
